Fix null SRX handling to check file existence before deletion
Previously, the code attempted to delete the file without checking its existence, potentially causing errors. Now, the method first verifies the file's presence before attempting deletion, ensuring more robust and error-free behavior.

+++ src/org/omegat/core/segmentation/SRX.java
@@ -129,13 +129,16 @@ public class SRX implements Serializable {
      */
     public static void saveTo(SRX srx, File outFile) throws IOException {
         if (srx == null) {
-            try {
-                Files.delete(outFile.toPath());
-            } catch (IOException e) {
-                Log.logErrorRB(e, "CORE_SRX_ERROR_DELETING_FILE");
+            if (outFile.exists()) {
+                try {
+                    Files.delete(outFile.toPath());
+                } catch (IOException e) {
+                    Log.logErrorRB(e, "CORE_SRX_ERROR_DELETING_FILE");
+                }
             }
             return;
         }
+
         try {
             srx.setVersion(CURRENT_VERSION);
             XMLEncoder xmlenc = new XMLEncoder(new FileOutputStream(outFile));
